requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.7
defusedxml==0.7.1
filelock==3.29.0
fsspec==2026.4.0
idna==3.11
Jinja2==3.1.6
llvmlite==0.47.0
MarkupSafe==3.0.3
more-itertools==11.0.2
mpmath==1.3.0
networkx==3.6.1
numba==0.65.1
numpy==2.4.4
openai-whisper==20250625
regex==2026.4.4
requests==2.33.1
setuptools==81.0.0
sympy==1.14.0
tiktoken==0.12.0
torch==2.11.0
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
youtube-transcript-api==1.2.4
yt-dlp==2026.3.17
